pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = "Telegram bot simple interface"
authors = ["Izbushko <izbaizbushka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "==3.*"
python-dotenv = "^1.0.0"
sqlmodel = "^0.0.14"
alembic = "^1.13.1"
pydantic-settings = "^2.1.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
